fix(tiny_fd): assert disconnection on (re-)connection request of a connected peer
This may happen for example if the peer was reset and right after boot
it sends the SABM or SNRM frame. If the startup time of the peer is
short (in the worst case it is enough that the startup time is less than
the keep-alive timeout), it is possible that the primary has not
detected the disconnection of peer yet.